feat: Refactor open-toko endpoint to remove duplicate check
The code changes refactor the open-toko endpoint in the User.routes.js file to remove a duplicate check for toko ownership. Previously, the endpoint was checking if the user already had a toko using the tokoMiddlewares.isAlreadyTokoOwner middleware, but this check was redundant as it was already being performed in the User.services.js file. The duplicate check has been removed to simplify the code and improve performance.